pyproject.toml
[project]
name = "ai-diagnostic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
